Terminando Configurations e iniciando Controller de teste

+++ PizzariaAPI/PizzariaAPI/Models/Cupom.cs
@@ -3,7 +3,7 @@
     public class Cupom
     {
         public int IdCupom { get; set; }
-        public float PorcentagemDesconto { get; set; }
+        public decimal PorcentagemDesconto { get; set; }
         public string CodigoCupom { get; set; }
         public bool Status{ get; set; }
     }
